Modifier méthode ajouterLigneTache Ajout d’une arrayList dans client des tâches qu’il propose

+++ src/main/java/Tache/AbstractLigneTache.java
@@ -13,13 +13,13 @@ public abstract class AbstractLigneTache {
     protected AbstractIntervenant intervenant;
     protected Integer codeLT;
     protected String libLT;
-    protected Float prixLT;
+    protected Double prixLT;
     protected String etatLT;
     protected Integer nbPersonne;
     protected AbstractCompetence competence;
     protected ArrayList<AbstractMessage> listMessage=new ArrayList<>();
     protected ArrayList<AbstractIntervenant> listIntervenant;
-    public AbstractLigneTache(AbstractCompetence competence,Integer id,Tache tache, String libLT, Float prixLT, String etatLT, Integer nbPersonne) {
+    public AbstractLigneTache(AbstractCompetence competence,Integer id,Tache tache, String libLT, Double prixLT, String etatLT, Integer nbPersonne) {
         this.competence=competence;
         this.libLT = libLT;
         this.prixLT = prixLT;
@@ -31,7 +31,7 @@ public abstract class AbstractLigneTache {
 
     }
 
-    public void setPrixLT(Float prixLT) {
+    public void setPrixLT(Double prixLT) {
         this.prixLT = prixLT;
     }
